Seção 9: DTOs & Data Annotations

+++ Server/src/ProEventos.API/Controllers/EventoController.cs
@@ -2,8 +2,8 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ProEventos.Application.DTO;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
 
 namespace ProEventos.API.Controllers
 {
@@ -23,7 +23,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado!");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -40,7 +40,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Nenhum evento encontrado!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -57,7 +57,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado!");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -70,7 +70,7 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
@@ -88,7 +88,7 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
@@ -110,10 +110,13 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var eventos = await _eventoService.GetEventoByIdAsync(id, false);
+                if (eventos == null) return NoContent();
+
                 var success = await _eventoService.Delete(id);
                 if (success) return Ok("Evento apagado!");
 
-                return BadRequest("Evento não deletado!");
+                throw new Exception("Ocorreu um erro não especificado!");
             }
             catch (Exception ex)
             {
